feat(sample) bring back old metisa.js
- update README
