+++ sample/browser/index.html
@@ -16,6 +16,18 @@
     <div>Sample Shop</div>
     <a href="./product-1.html">Product 1</a>
 
+    <div class="mt-widget" data-widget-id=1></div>
+
+    <script
+      src="https://code.jquery.com/jquery-3.2.1.min.js"
+      integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4="
+      crossorigin="anonymous"></script>
+
     <script src="../../dist/browser.js"></script>
+
+    <script type="text/javascript">
+      mt('store', 'metisa');
+      mt('token', 'demo-token');
+    </script>
   </body>
 </html>
