+++ sample/browser/product-1.html
@@ -15,6 +15,10 @@
   <body>
     <div>Sample Product 1</div>
 
-    <script src='../../dist/browser.js' ></script>
+    <script
+      src="https://code.jquery.com/jquery-3.2.1.min.js"
+      integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4="
+      crossorigin="anonymous"></script>
+    <script src="../../dist/browser.js"></script>
   </body>
 </html>
