+++ src/Metisa/dom/withIFrame.js
@@ -1,0 +1,215 @@
+module.exports = function(composedClass) {
+  return class IFrame extends composedClass {
+    createIFrameWithId(id) {
+      var iframe = document.createElement('iframe');
+
+      iframe.setAttribute('style', 'border: 0px; width: 100%;');
+      iframe.setAttribute('frameborder', '0');
+      iframe.setAttribute('scrolling', 'no');
+      iframe.setAttribute('onload', 'window.Metisa.resizeIframe(this)');
+      iframe.setAttribute('id', 'widget-' + id);
+
+      return iframe;
+    }
+
+    resizeIframe(obj) {
+      obj.style.height = 0;
+      obj.style.height = obj.contentWindow.document.body.offsetHeight + 'px';
+    }
+
+    getLoaderHTML() {
+      return '<!doctype html><html><style>body{height: 100px;}.cs-loader{height: 100%; width: 100%;}.cs-loader-inner{transform: translateY(-50%); top: 50%; position: absolute; width: calc(100% - 200px); color: #A2A3A3; padding: 0 100px; text-align: center;}.cs-loader-inner label{font-size: 20px; opacity: 0; display: inline-block;}@keyframes lol{0%{opacity: 0; transform: translateX(-300px);}33%{opacity: 1; transform: translateX(0px);}66%{opacity: 1; transform: translateX(0px);}100%{opacity: 0; transform: translateX(300px);}}@-webkit-keyframes lol{0%{opacity: 0; -webkit-transform: translateX(-300px);}33%{opacity: 1; -webkit-transform: translateX(0px);}66%{opacity: 1; -webkit-transform: translateX(0px);}100%{opacity: 0; -webkit-transform: translateX(300px);}}.cs-loader-inner label:nth-child(6){-webkit-animation: lol 3s infinite ease-in-out; animation: lol 3s infinite ease-in-out;}.cs-loader-inner label:nth-child(5){-webkit-animation: lol 3s 100ms infinite ease-in-out; animation: lol 3s 100ms infinite ease-in-out;}.cs-loader-inner label:nth-child(4){-webkit-animation: lol 3s 200ms infinite ease-in-out; animation: lol 3s 200ms infinite ease-in-out;}.cs-loader-inner label:nth-child(3){-webkit-animation: lol 3s 300ms infinite ease-in-out; animation: lol 3s 300ms infinite ease-in-out;}.cs-loader-inner label:nth-child(2){-webkit-animation: lol 3s 400ms infinite ease-in-out; animation: lol 3s 400ms infinite ease-in-out;}.cs-loader-inner label:nth-child(1){-webkit-animation: lol 3s 500ms infinite ease-in-out; animation: lol 3s 500ms infinite ease-in-out;}</style><body> <div class="cs-loader"> <div class="cs-loader-inner"> <label>●</label> <label>●</label> <label>●</label> <label>●</label> <label>●</label> <label>●</label> </div></div></body></html>';
+    }
+
+    decodeHtmlEntities(str) {
+      return str.replace(/&#?(\w+);/g, function(match, dec) {
+        if (isNaN(dec)) {
+          var chars = {
+              quot: 34,
+              amp: 38,
+              lt: 60,
+              gt: 62,
+              nbsp: 160,
+              copy: 169,
+              reg: 174,
+              deg: 176,
+              frasl: 47,
+              trade: 8482,
+              euro: 8364,
+              Agrave: 192,
+              Aacute: 193,
+              Acirc: 194,
+              Atilde: 195,
+              Auml: 196,
+              Aring: 197,
+              AElig: 198,
+              Ccedil: 199,
+              Egrave: 200,
+              Eacute: 201,
+              Ecirc: 202,
+              Euml: 203,
+              Igrave: 204,
+              Iacute: 205,
+              Icirc: 206,
+              Iuml: 207,
+              ETH: 208,
+              Ntilde: 209,
+              Ograve: 210,
+              Oacute: 211,
+              Ocirc: 212,
+              Otilde: 213,
+              Ouml: 214,
+              times: 215,
+              Oslash: 216,
+              Ugrave: 217,
+              Uacute: 218,
+              Ucirc: 219,
+              Uuml: 220,
+              Yacute: 221,
+              THORN: 222,
+              szlig: 223,
+              agrave: 224,
+              aacute: 225,
+              acirc: 226,
+              atilde: 227,
+              auml: 228,
+              aring: 229,
+              aelig: 230,
+              ccedil: 231,
+              egrave: 232,
+              eacute: 233,
+              ecirc: 234,
+              euml: 235,
+              igrave: 236,
+              iacute: 237,
+              icirc: 238,
+              iuml: 239,
+              eth: 240,
+              ntilde: 241,
+              ograve: 242,
+              oacute: 243,
+              ocirc: 244,
+              otilde: 245,
+              ouml: 246,
+              divide: 247,
+              oslash: 248,
+              ugrave: 249,
+              uacute: 250,
+              ucirc: 251,
+              uuml: 252,
+              yacute: 253,
+              thorn: 254,
+              yuml: 255,
+              lsquo: 8216,
+              rsquo: 8217,
+              sbquo: 8218,
+              ldquo: 8220,
+              rdquo: 8221,
+              bdquo: 8222,
+              dagger: 8224,
+              Dagger: 8225,
+              permil: 8240,
+              lsaquo: 8249,
+              rsaquo: 8250,
+              spades: 9824,
+              clubs: 9827,
+              hearts: 9829,
+              diams: 9830,
+              oline: 8254,
+              larr: 8592,
+              uarr: 8593,
+              rarr: 8594,
+              darr: 8595,
+              hellip: 133,
+              ndash: 150,
+              mdash: 151,
+              iexcl: 161,
+              cent: 162,
+              pound: 163,
+              curren: 164,
+              yen: 165,
+              brvbar: 166,
+              brkbar: 166,
+              sect: 167,
+              uml: 168,
+              die: 168,
+              ordf: 170,
+              laquo: 171,
+              not: 172,
+              shy: 173,
+              macr: 175,
+              hibar: 175,
+              plusmn: 177,
+              sup2: 178,
+              sup3: 179,
+              acute: 180,
+              micro: 181,
+              para: 182,
+              middot: 183,
+              cedil: 184,
+              sup1: 185,
+              ordm: 186,
+              raquo: 187,
+              frac14: 188,
+              frac12: 189,
+              frac34: 190,
+              iquest: 191,
+              Alpha: 913,
+              alpha: 945,
+              Beta: 914,
+              beta: 946,
+              Gamma: 915,
+              gamma: 947,
+              Delta: 916,
+              delta: 948,
+              Epsilon: 917,
+              epsilon: 949,
+              Zeta: 918,
+              zeta: 950,
+              Eta: 919,
+              eta: 951,
+              Theta: 920,
+              theta: 952,
+              Iota: 921,
+              iota: 953,
+              Kappa: 922,
+              kappa: 954,
+              Lambda: 923,
+              lambda: 955,
+              Mu: 924,
+              mu: 956,
+              Nu: 925,
+              nu: 957,
+              Xi: 926,
+              xi: 958,
+              Omicron: 927,
+              omicron: 959,
+              Pi: 928,
+              pi: 960,
+              Rho: 929,
+              rho: 961,
+              Sigma: 931,
+              sigma: 963,
+              Tau: 932,
+              tau: 964,
+              Upsilon: 933,
+              upsilon: 965,
+              Phi: 934,
+              phi: 966,
+              Chi: 935,
+              chi: 967,
+              Psi: 936,
+              psi: 968,
+              Omega: 937,
+              omega: 969
+          };
+
+          if (chars[dec] !== undefined) {
+            dec = chars[dec];
+          }
+        }
+        return String.fromCharCode(dec);
+      });
+    }
+  };
+};
