+++ src/browser.js
@@ -2,6 +2,8 @@ const MetisaDom = require('./Metisa/dom');
 const util = require('./util');
 
 module.exports = (function() {
-  if (util.environment !== 'browser' ) { return console.warn('Metisa browser can only run inside a browser'); }
+  if (util.environment !== 'browser' ) {
+    return console.warn('Metisa browser can only run inside a browser');
+  }
   window.Metisa = new MetisaDom();
 })();
